requirements.txt
Django==1.5.1
South==0.8.1
django-extensions==1.1.1
django-tastypie==0.9.15
hammock==0.2.4
ipython==0.13.2
mimeparse==0.1.3
python-dateutil==2.1
pyzmq==13.1.0
requests==1.2.3
six==1.3.0
tornado==3.0.1
wsgiref==0.1.2
